Move order views urls to separate file
